+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 1.0.0 (Jan 1, 2014)
 
 * Initial release
+
+## 1.1.0 (May 4, 2023)
+
+* Create EscalationLogger class to all for automatic escalation based off of log messages.
+* Create TimeBasedEscalationTrigger to use with EscalationLogger.